refactor: modularize upgrade-system contract by splitting logic into separate modules for storage, types, errors, and implementation.
